Changed return types for grouping
Before, a key which returned no results would still be there but with a
value of `null`, or in some cases an empty list. This is wasteful for
transmissions, makes it easier for DoS, and takes an unnecessary toll on
performance. Now just returns key for which a value was found